+++ example/webpack.config.js
@@ -4,11 +4,15 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 const webpack = require('webpack')
 
-const HOST = '0.0.0.0'
+const HOST = 'localhost'
 const PORT = 8081
 
 const getPublicPath = () => {
-  if (!!process.env.NETLIFY || process.env.NODE_ENV !== 'production') {
+  if (
+    process.env.NETLIFY ||
+    process.env.SURGE ||
+    process.env.NODE_ENV !== 'production'
+  ) {
     return '/'
   }
   return '/react-music-player'