+++ src/components/PlayerMobile.js
@@ -30,6 +30,7 @@ const PlayerMobile = ({
   locale,
   toggleMode,
   renderAudioTitle,
+  shouldShowPlayIcon,
 }) => (
   <div className={cls(prefix, { 'default-bg': !glassBg, 'glass-bg': glassBg })}>
     <PlayModeTip
@@ -87,10 +88,14 @@ const PlayerMobile = ({
       ) : (
         <span
           className="group play-btn"
-          title={playing ? locale.clickToPauseText : locale.clickToPlayText}
+          title={
+            shouldShowPlayIcon
+              ? locale.clickToPlayText
+              : locale.clickToPauseText
+          }
           onClick={onPlay}
         >
-          {playing ? icon.pause : icon.play}
+          {shouldShowPlayIcon ? icon.play : icon.pause}
         </span>
       )}
       <span