+++ src/utils.js
@@ -1,12 +1,14 @@
-//秒转换成 时间格式
+/* eslint-disable no-param-reassign */
+/* eslint-disable radix */
+/* eslint-disable no-bitwise */
 export function formatTime(second) {
   let i = 0
   let h = 0
   let s = parseInt(second)
-  if (s > 60) {
+  if (s >= 60) {
     i = parseInt(s / 60)
     s = parseInt(s % 60)
-    if (i > 60) {
+    if (i >= 60) {
       h = parseInt(i / 60)
       i = parseInt(i % 60)
     }
@@ -14,7 +16,7 @@ export function formatTime(second) {
   // 补零
   const zero = (v) => (v >> 0 < 10 ? `0${v}` : v)
   if (h > 0) return [zero(h), zero(i), zero(s)].join(':')
-  else return [zero(i), zero(s)].join(':')
+  return [zero(i), zero(s)].join(':')
 }
 
 export function createRandomNum(minNum, maxNum) {
@@ -24,7 +26,7 @@ export function createRandomNum(minNum, maxNum) {
 export function distinct(array) {
   return array
     .map((item) => JSON.stringify(item))
-    .filter((item, idx, arry) => idx === arry.indexOf(item))
+    .filter((item, idx, arr) => idx === arr.indexOf(item))
     .map((item) => JSON.parse(item))
 }
 
@@ -43,4 +45,68 @@ export const isSafari = () => {
   return (
     /Safari/.test(navigator.userAgent) && !/Chrome/.test(navigator.userAgent)
   )
+}
+
+// https://stackoverflow.com/a/9039885/2789451
+export function isIOS() {
+  return (
+    [
+      'iPad Simulator',
+      'iPhone Simulator',
+      'iPod Simulator',
+      'iPad',
+      'iPhone',
+      'iPod',
+    ].includes(navigator.platform) ||
+    // iPad on iOS 13 detection
+    (navigator.userAgent.includes('Mac') && 'ontouchend' in document)
+  )
+}
+
+// https://stackoverflow.com/questions/7451508/html5-audio-playback-with-fade-in-and-fade-out
+export function swing(p) {
+  return 0.5 - Math.cos(p * Math.PI) / 2
+}
+
+export function adjustVolume(
+  element,
+  startVolume,
+  endVolume,
+  { duration = 1000, easing = swing, interval = 13 } = {},
+  callback,
+) {
+  let delta = endVolume - startVolume
+
+  if (!delta || !duration || !easing || !interval || isIOS()) {
+    element.volume = endVolume
+    callback()
+    return { fadeInterval: undefined, updateIntervalEndVolume: undefined }
+  }
+
+  const ticks = Math.floor(duration / interval)
+  let tick = 1
+
+  const updateIntervalEndVolume = (newVolume) => {
+    endVolume = newVolume
+  }
+
+  const timer = setInterval(() => {
+    // End volume may have changed in middle of fading
+    const newDelta = endVolume - startVolume
+    if (newDelta !== delta) {
+      delta = newDelta
+    }
+
+    element.volume = startVolume + easing(tick / ticks) * delta
+    if (++tick >= ticks) {
+      element.volume = endVolume
+      clearInterval(timer)
+      callback()
+    }
+  }, interval)
+
+  return {
+    fadeInterval: timer,
+    updateIntervalEndVolume,
+  }
 }