+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+## [4.21.2](https://github.com/lijinke666/react-music-player/compare/v4.21.1...v4.21.2) (2021-01-27)
+
+
+### Bug Fixes
+
+* cannot reset play index when audioLists changed [#286](https://github.com/lijinke666/react-music-player/issues/286) ([e42177b](https://github.com/lijinke666/react-music-player/commit/e42177b187bd0a56914d0c59d0db2113fd694867))
+* optimize progress bar seek step [#283](https://github.com/lijinke666/react-music-player/issues/283) ([fbff1e3](https://github.com/lijinke666/react-music-player/commit/fbff1e35ace2695a102c3cdfa08a44a424119b7f))
+
+## [4.21.1](https://github.com/lijinke666/react-music-player/compare/v4.21.0...v4.21.1) (2021-01-15)
+
+
+### Bug Fixes
+
+* do not fade on iOS devices ([767df65](https://github.com/lijinke666/react-music-player/commit/767df659d9f111e5097855f426d2a05b32f6f08e))
+* no longer mutes on play when fadeIn is 0 ([6cab882](https://github.com/lijinke666/react-music-player/commit/6cab8823db679cfdcb86572af1bce53af0ec38a4))
+
+# [4.21.0](https://github.com/lijinke666/react-music-player/compare/v4.20.1...v4.21.0) (2021-01-13)
+
+
+### Bug Fixes
+
+* auto play when audio seeking by progress bar [#260](https://github.com/lijinke666/react-music-player/issues/260) ([016af3b](https://github.com/lijinke666/react-music-player/commit/016af3b7793b648fde7e3e8ed849dece39876c68))
+* avoid play/pause button re-render by setting playing to false earlier ([2568f31](https://github.com/lijinke666/react-music-player/commit/2568f31e325e139c65b4b92de1e206bcb2ebc26d))
+* can not trigger onAudioVolumeChange when play or pause audio with volume fade mode ([6b88d6c](https://github.com/lijinke666/react-music-player/commit/6b88d6c7ad3a113ee4bf62535558c46dd3ac23ef))
+* play button delay in mobile player ([5e8eccd](https://github.com/lijinke666/react-music-player/commit/5e8eccdf15eb89c086a051eb32903e9bccfba97b))
+* set audio volume to pre-fade volume before fading in ([63c6b6f](https://github.com/lijinke666/react-music-player/commit/63c6b6fe57575a288d44b2584fc7b1e64a8eb5c0))
+
+
+### Features
+
+* implement non-linear volume slider ([a7b9cec](https://github.com/lijinke666/react-music-player/commit/a7b9cec7168b4dec798d489aaa843207628e8c11))
+
+## [4.20.1](https://github.com/lijinke666/react-music-player/compare/v4.20.0...v4.20.1) (2021-01-11)
+
+
+### Bug Fixes
+
+* volume fade-in and fade-out issues [#256](https://github.com/lijinke666/react-music-player/issues/256) ([9ab8b07](https://github.com/lijinke666/react-music-player/commit/9ab8b07b070fbd21dc4ee648eec54f0bf17c92ed))
+
+# [4.20.0](https://github.com/lijinke666/react-music-player/compare/v4.19.1...v4.20.0) (2021-01-05)
+
+
+### Bug Fixes
+
+* error timestamp format [#242](https://github.com/lijinke666/react-music-player/issues/242) ([6ecbf73](https://github.com/lijinke666/react-music-player/commit/6ecbf737f6c5d8f3accdaca1f8542b2f20d4685e))
+* invalid audio volume fadeIn [#238](https://github.com/lijinke666/react-music-player/issues/238) ([1120e28](https://github.com/lijinke666/react-music-player/commit/1120e286c6ca4e645b07e293ce59dd004a941ddc))
+
+
+### Features
+
+* support audio volume fadeIn and fadeOut [#238](https://github.com/lijinke666/react-music-player/issues/238) ([dedd22c](https://github.com/lijinke666/react-music-player/commit/dedd22c73b1415e8c3f5fcedcd00e97637610652))
+
 ## [4.19.1](https://github.com/lijinke666/react-music-player/compare/v4.19.0...v4.19.1) (2020-11-07)